+++ lib/src/internal/content_decoder/fallback_decoder.dart
@@ -1,4 +1,9 @@
-part of 'response_decode.dart';
+import 'dart:async';
+
+import 'package:http/http.dart' show Response;
+import 'package:rest_client/src/response.dart';
+
+import 'response_decode_strategy.dart';
 
 final class FallbackDecodeStrategy extends ResponseDecodeStrategy {
   const FallbackDecodeStrategy();