+++ lib/src/internal/rest_methods_interface.dart
@@ -1,33 +1,42 @@
 part of '../../rest_client.dart';
 
 abstract interface class _IRestMethods {
-  Future<RCResponse> get(String path, {Map<String, dynamic>? queryParameters, Map<String, String>? headers});
-
-  Future<RCResponse> post(String path,
-      {RequestBody? body, Map<String, dynamic>? queryParameters, Map<String, String>? headers});
-
-  Future<RCResponse> put(String path,
-      {RequestBody? body, Map<String, dynamic>? queryParameters, Map<String, String>? headers});
-
-  Future<RCResponse> delete(String path,
-      {RequestBody? body, Map<String, dynamic>? queryParameters, Map<String, String>? headers});
-
-  Future<RCResponse> patch(String path,
-      {RequestBody? body, Map<String, dynamic>? queryParameters, Map<String, String>? headers});
-
-  Future<RCResponse> sendImage(String path,
-      {required BytesBody bytes, Map<String, dynamic>? queryParameters, Map<String, String>? headers});
-
-  Future<RCResponse> sendFile(String path,
-      {required String title,
-      required BytesBody bytes,
-      Map<String, dynamic>? queryParameters,
-      Map<String, String>? headers});
-
-  ///Experimental download file stream
+  /// {@template rest_client.http_methods.send}
+  /// Base send request method.
   ///
-  ///'Download' is progress
-  ///base64 String is completed
-  Stream<DownloadFile> downloadFile(String api, Map<String, dynamic> body,
-      {Map<String, String> queryParameters = const {}});
+  /// Used [HttpMethod] for create [request].
+  ///
+  /// Example:
+  /// ```dart
+  ///
+  /// send(Request(HttpMethod.get, Uri.tryParse('example.com')))
+  ///
+  /// ```
+  /// {@endtemplate}
+  Future<StreamedResponse> send(BaseRequest request, {final RequestBody? body, final Map<String, String>? headers});
+
+  /// {@template rest_client.http_methods.get}
+  /// GET method.
+  /// {@endtemplate}
+  Future<RCResponse> get(Uri uri, {Map<String, String>? headers});
+
+  /// {@template rest_client.http_methods.post}
+  /// POST method.
+  /// {@endtemplate}
+  Future<RCResponse> post(Uri uri, {RequestBody? body, Map<String, String>? headers});
+
+  /// {@template rest_client.http_methods.put}
+  /// PUT method.
+  /// {@endtemplate}
+  Future<RCResponse> put(Uri uri, {RequestBody? body, Map<String, String>? headers});
+
+  /// {@template rest_client.http_methods.delete}
+  /// DELETE method.
+  /// {@endtemplate}
+  Future<RCResponse> delete(Uri uri, {RequestBody? body, Map<String, String>? headers});
+
+  /// {@template rest_client.http_methods.patch}
+  /// PATCH method.
+  /// {@endtemplate}
+  Future<RCResponse> patch(Uri uri, {RequestBody? body, Map<String, String>? headers});
 }