+++ lib/src/session/jwt_store.dart
@@ -1,9 +1,0 @@
-part of '../../rest_client.dart';
-
-abstract interface class IJWTStore {
-  const IJWTStore();
-
-  Future<JWT?> read();
-  FutureOr<void> write(JWT value);
-  FutureOr<void> remove();
-}