+++ lib/src/utils/json_parser/js.dart
@@ -1,15 +1,22 @@
 import 'dart:async';
+import 'dart:convert' show JsonCodec;
 
 import 'json_parser.dart';
 
-final class WebJsonParser with JsonCodecMixin implements JsonParser {
+/// {@nodoc}
+JsonParser getParser() => WebJsonParser();
+
+/// {@nodoc}
+final class WebJsonParser implements JsonParser {
   static const WebJsonParser _internalSingleton = WebJsonParser._internal();
   factory WebJsonParser() => _internalSingleton;
   const WebJsonParser._internal();
 
+  final JsonCodec _jsonCodec = JsonParser.jsonCodec;
+
   @override
   @pragma('dart2js:tryInline')
-  Future<Map<String, Object?>> decode(String data, {bool useIsolate = false, String? debugPrint}) async {
+  Future<Map<String, Object?>> strDecodeJson(String data, {bool useIsolate = false, String? debugPrint}) async {
     // From flutter/foundation/_isolates_web.dart
     //
     // To avoid blocking the UI immediately for an expensive function call, we
@@ -17,12 +24,12 @@ final class WebJsonParser with JsonCodecMixin implements JsonParser {
     // of work.
 
     await Future<void>.delayed(Duration.zero);
-    return decode(data);
+    return _jsonCodec.decode(data) as Map<String, Object?>;
   }
 
   @override
   @pragma('dart2js:tryInline')
-  Future<String> encode(Map<String, Object?> data, {bool useIsolate = false, String? debugPrint}) async {
+  Future<String> objEncode(Map<String, Object?> data, {bool useIsolate = false, String? debugPrint}) async {
     // From flutter/foundation/_isolates_web.dart
     //
     // To avoid blocking the UI immediately for an expensive function call, we
@@ -30,8 +37,6 @@ final class WebJsonParser with JsonCodecMixin implements JsonParser {
     // of work.
 
     await Future<void>.delayed(Duration.zero);
-    return encodeObject(data);
+    return _jsonCodec.encode(data);
   }
 }
-
-JsonParser getParser() => WebJsonParser();