+++ CHANGELOG.md
@@ -6,3 +6,11 @@
 
 * Json utils change
 * Structure project change
+
+# 0.0.3
+
+* Added Interceptor API.
+* Remove url from [config](lib/src/config.dart). Use Uri API.
+* Remove queryParams method signature. Use Uri API.
+* Used [lints](https://pub.dev/packages/lints) package.
+* [WIP] Doc